+++ CHANGELOG.md
@@ -4,10 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [2.7.0] - 2020-11-20 - [#115](https://github.com/mza921/Plex-Auto-Collections/pull/115)
+## [2.7.0] - 2020-11-24 - [#115](https://github.com/mza921/Plex-Auto-Collections/pull/115)
 ### Added
 - Added `tmdb_trending_daily` and `tmdb_trending_weekly`
 - [#98](https://github.com/mza921/Plex-Auto-Collections/issues/98) - Added `cache` attribute to cache the IDs of movies/shows for quicker lookup
+- [#123](https://github.com/mza921/Plex-Auto-Collections/issues/123) - Added new filter `plex_collection`
 
 ### Changed
 - Created a mapping for TMDb ID to Plex Rating Key once each run instead of every time tmdb, imdb, tvdb, or trakt list was run.